Move AdvisoryDatabase under the ::db module
Re-exported at the toplevel for easier use and backwards compatibility
